Small ideas on how to provide a Context of a domain. Sky is the limit :rocket:
