+++ lib/remit/comments/ingest_comments/ingest_comment_test.ex
@@ -1,0 +1,3 @@
+defmodule Remit.IngestComment.Test do
+  # Tests..
+end
